+++ blueprints/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "blueprints.apps.BlueprintsConfig"
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"
 __title__ = "Blueprints"