+++ blueprints/auth_hooks.py
@@ -4,6 +4,7 @@ from allianceauth import hooks
 from allianceauth.services.hooks import MenuItemHook, UrlHook
 
 from . import urls
+from .models import Request
 
 
 class BlueprintLibraryMenuItem(MenuItemHook):
@@ -21,6 +22,13 @@ class BlueprintLibraryMenuItem(MenuItemHook):
 
     def render(self, request):
         if request.user.has_perm("blueprints.basic_access"):
+            if request.user.has_perm("blueprints.manage_requests"):
+                app_count = Request.objects.requests_fulfillable_by_user(
+                    request.user
+                ).count()
+            else:
+                app_count = None
+            self.count = app_count if app_count and app_count > 0 else None
             return MenuItemHook.render(self, request)
         return ""