+++ blueprints/static/blueprints/js/blueprint_list.js
@@ -3,12 +3,15 @@
 $(document).ready(function () {
     "use strict";
 
-    var listDataUrl = blueprintsDataTableSettings.listDataurl;
+    var listDataUrl = blueprintsDataTableSettings.listDataUrl;
+    var createRequestUrl = blueprintsDataTableSettings.createRequestUrl;
+    var createRequestModalUrl =
+        blueprintsDataTableSettings.createRequestModalUrl;
     var dataTablesPageLength = blueprintsDataTableSettings.dataTablesPageLength;
     var dataTablesPaging = blueprintsDataTableSettings.dataTablesPaging;
-
+    var csrfToken = blueprintsDataTableSettings.csrfToken;
     /* dataTable def */
-    $("#tab_blueprints").DataTable({
+    $("#table-blueprints").DataTable({
         ajax: {
             url: listDataUrl,
             dataSrc: "",
@@ -17,14 +20,17 @@ $(document).ready(function () {
 
         columns: [
             { data: "type_icon" },
-            { data: "type" },
-            { data: "quantity" },
+            { data: "type_name" },
+            {
+                className: "right-column",
+                data: "quantity",
+            },
             { data: "owner" },
             { data: "material_efficiency" },
             { data: "time_efficiency" },
             { data: "original" },
             { data: "runs" },
-
+            { data: "blueprint_id" },
             // hidden columns
             { data: "location" },
             { data: "filter_is_original" },
@@ -40,19 +46,39 @@ $(document).ready(function () {
         pageLength: dataTablesPageLength,
 
         columnDefs: [
-            { sortable: false, targets: [0, 2] },
-            { visible: false, targets: [8, 9] },
+            { sortable: false, targets: [0, 2, 8] },
+            { visible: false, targets: [9, 10] },
+            {
+                render: function (data, type, row) {
+                    if (type === "display") {
+                        if (row["original"] != "") {
+                            return (
+                                '<button class="btn btn-success" data-toggle="modal" data-target="#modalCreateRequestContainer" data-ajax_url="' +
+                                createRequestModalUrl +
+                                "?blueprint_id=" +
+                                data +
+                                '" aria-label="Create Request"><span class="fas fa-copy"></span></button>'
+                            );
+                        } else {
+                            return "";
+                        }
+                    }
+
+                    return data;
+                },
+                targets: [8],
+            },
         ],
 
         order: [
-            [8, "asc"],
+            [9, "asc"],
             [1, "asc"],
         ],
 
         filterDropDown: {
             columns: [
                 {
-                    idx: 8,
+                    idx: 9,
                     title:
                         blueprintsDataTableSettings.translation.filterLocation,
                 },
@@ -60,7 +86,7 @@ $(document).ready(function () {
                 { idx: 4 },
                 { idx: 5 },
                 {
-                    idx: 9,
+                    idx: 10,
                     title:
                         blueprintsDataTableSettings.translation
                             .filterIsOriginal,
@@ -75,14 +101,14 @@ $(document).ready(function () {
             var rows = api.rows({ page: "current" }).nodes();
             var last = null;
 
-            api.column(8, { page: "current" })
+            api.column(9, { page: "current" })
                 .data()
                 .each(function (group, i) {
                     if (last !== group) {
                         $(rows)
                             .eq(i)
                             .before(
-                                '<tr class="tr-group"><td colspan="8">' +
+                                '<tr class="tr-group"><td colspan="9">' +
                                     group +
                                     "</td></tr>"
                             );