+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.0 - 2020-12-17
+
+### Added
+
+- Ability to request blueprint copies
+
 ## 0.1.1 - 2020-12-14
 
 ### Fixed
 
-- Samll bug in Owner object
+- Small bug in Owner object
 
 ## 0.1.0 - 2020-12-14